test: move codec tests to a separate file
